App routing+using queries pt1 - not working
