+++ apps/frontend/src/api/hooks/eventQueryHooks.ts
@@ -1,7 +1,8 @@
+'use client';
 import { useQuery } from '@tanstack/react-query';
 import axios from 'axios';
 
-import { EventDetails } from '@/pages/event/types/eventDetails';
+import { EventDetails } from '@/app/events/types/eventDetails';
 
 const url = 'http://localhost:3001';
