+++ apps/frontend/src/app/events/page.tsx
@@ -1,0 +1,3 @@
+export default function EventList() {
+  return <p>Events</p>;
+}
