+++ apps/frontend/src/app/events/types/eventDetails.ts
@@ -1,5 +1,5 @@
 import { EventModel } from '@/api/model/event.model';
-import { PublicUser } from '@/pages/user/types/PublicUser';
+import { PublicUser } from '@/app/users/types/PublicUser';
 
 export type EventDetails = EventModel & {
   owner: PublicUser;
