+++ apps/frontend/src/app/users/types/PublicUser.ts
