+++ apps/frontend/src/util/environment.ts
@@ -1,0 +1,2 @@
+export const API_HOST = 'http://localhost:3001';
+export const FRONTEND_HOST = 'http://localhost:3000';
